# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.10",
    "title": "javaee7-angular"
  },
  "host": "localhost:8080",
  "basePath": "/javaee7-angular/resources",
  "schemes": [
    "http"
  ],
  "paths": {
    "/persons": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "page",
            "in": "query",
            "required": false,
            "default": "1"
          },
          {
            "type": "string",
            "name": "sortDirections",
            "in": "query",
            "required": false,
            "default": "asc"
          },
          {
            "type": "string",
            "name": "sortFields",
            "in": "query",
            "required": false,
            "default": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/PaginatedListWrapper"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        }
      }
    },
    "/persons/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {}
          }
        }
      }
    },
    "/tables": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Table"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/Table"
            }
          }
        }
      }
    },
    "/tables/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/Table"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {}
          }
        }
      }
    },
    "/restaurants": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "page",
            "in": "query",
            "required": false,
            "default": "1"
          },
          {
            "type": "string",
            "name": "sortDirections",
            "in": "query",
            "required": false,
            "default": "asc"
          },
          {
            "type": "string",
            "name": "sortFields",
            "in": "query",
            "required": false,
            "default": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/PaginatedListWrapper"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR",
            "headers": {}
          }
        }
      }
    },
    "/restaurants/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "headers": {}
          },
          "500": {
            "description": "INTERNAL SERVER ERROR",
            "headers": {}
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {}
          },
          "500": {
            "description": "INTERNAL SERVER ERROR",
            "headers": {}
          }
        }
      }
    }
  },
  "definitions": {
    "Object": {
      "properties": {}
    },
    "PaginatedListWrapper": {
      "properties": {
        "currentPage": {
          "type": "integer"
        },
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Object"
          }
        },
        "pageSize": {
          "type": "integer"
        },
        "sortDirections": {
          "type": "string"
        },
        "sortFields": {
          "type": "string"
        },
        "totalResults": {
          "type": "integer"
        }
      }
    },
    "Person": {
      "properties": {
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "imageUrl": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Table": {
      "properties": {
        "number": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "available_places": {
          "type": "integer"
        },
        "movable": {
          "type": "boolean"
        }
      }
    },
    "Restaurant": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "latitude": {
          "type": "number"
        },
        "longitude": {
          "type": "number"
        },
        "speciality": {
          "type": "string"
        },
        "url_img": {
          "type": "string"
        },
        "tel_number": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    }
  }
}